Latest renderer commit.
Includes another activity which can attempt to generate tiles if given the tile coords (zoom/x/y).

Currently there are problems when querying the database in order to fetch the list of features which should be generated.

Also included in this commit is a bit of a clean-up to remove any warnings in a few classes.