Fixed null-pointer exception and a few other issues (off-by-one error).
Also added a destructor to handle the cleanup operation for sqlite database handles.

Revert a change accidently committed to SampleExtensive which made the internal tile provider the default. There is now a sample tile renderer which can be used to test the internal render instead.

SQL performance is actually OK if the correct indexes are in place - all SQL queries complete within a few seconds now. Just need to do further debugging on the style-filters to figure out why nothing is actually being draw.